ICP as an independent class
