+++ include/merge_pc2/merge_pc2.h
@@ -2,40 +2,37 @@
 // Created by Eason Hua on 5/2/24.
 //
 
-#ifndef MERGE_PCL_H
-#define MERGE_PCL_H
+#ifndef MERGE_PC2_H
+#define MERGE_PC2_H
 
 #include <thread>
 #include <deque>
 #include <mutex>
 #include <vector>
 #include <exception>
-
 #include <ros/ros.h>
-#include <ros/rate.h>
-
 #include <sensor_msgs/PointCloud2.h>
 #include <pcl_ros/transforms.h>
 #include <pcl_ros/point_cloud.h>
 #include <pcl/point_types.h>
 #include <pcl/kdtree/kdtree_flann.h>
-#include <pcl/registration/icp.h>
 #include <pcl_conversions/pcl_conversions.h>
-
 #include <tf/transform_listener.h>
+
+#include "merge_pc2/icp_utils.h"
 
 typedef pcl::PointXYZ PointT;
 typedef pcl::PointCloud<PointT> PointCloudT;
 
 using namespace std;
 
-class PointCloudMerger {
+class mergePC2 {
 public:
-    PointCloudMerger(ros::NodeHandle nh);
-    ~PointCloudMerger() = default;
+    mergePC2(ros::NodeHandle nh);
+    ~mergePC2() = default;
 
     // Method to process buffers
-    void processBuffers();
+    void mainLoop();
 
 private:
     // Subscribers
@@ -49,12 +46,7 @@ private:
     double duration;
     tf::TransformListener tf_listener;
 
-    // The Iterative Closest Point algorithm
-    bool icp_enable;
-    int icp_max_iter;
-    double icp_tf_epsilon;
-    double icp_euclidean_fit_epsilo;
-    double icp_max_coor_d;
+    icpUtils icp_utils;
 
     // Buffers to store point clouds
     std::deque<PointCloudT::Ptr> pcT_buffer;
@@ -67,7 +59,6 @@ private:
 
     // Unified callback for point cloud subscribers
     void pc2Callback(const sensor_msgs::PointCloud2ConstPtr& pc2, const size_t index);
-    void icpAlgorithm(PointCloudT::Ptr& pcT_out, const PointCloudT::Ptr& pcT);
 };
 
-#endif //MERGE_PCL_H
+#endif // MERGE_PC2_H
