+++ package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <package format="2">
     <name>easondrone_mapping</name>
-    <version>2.1.6</version>
+    <version>2.1.7</version>
     <description>
-        A ROS package to process PointCloud and build OctoMap
+        A ROS package to process PointCloud2 and build OctoMap
     </description>
 
     <maintainer email="12010508@mail.sustech.edu.cn">Eason Hua</maintainer>
